requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]
